removed settings layout
